// src/generated/models/Pm_skillsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_skillspm_skillcategory = {
  0: 'Functional',
  1: 'Technical',
  2: 'Certification'
} as const;
export type Pm_skillspm_skillcategory = keyof typeof Pm_skillspm_skillcategory;
export const Pm_skillsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_skillsstatecode = keyof typeof Pm_skillsstatecode;
export const Pm_skillsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_skillsstatuscode = keyof typeof Pm_skillsstatuscode;

export interface Pm_skillsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_isactive?: boolean;
  pm_skillcategory?: Pm_skillspm_skillcategory;
  pm_skilldescription?: string;
  pm_skillid: string;
  pm_skillname?: string;
  statecode: Pm_skillsstatecode;
  statuscode?: Pm_skillsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_skills extends Pm_skillsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_isactivename?: string;
  pm_skillcategoryname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}