// src/generated/models/Pm_resourceskillsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_resourceskillspm_proficiencylevel = {
  0: '_5_Expert',
  1: '_4_Proficient',
  2: '_3_Competent'
} as const;
export type Pm_resourceskillspm_proficiencylevel = keyof typeof Pm_resourceskillspm_proficiencylevel;
export const Pm_resourceskillsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_resourceskillsstatecode = keyof typeof Pm_resourceskillsstatecode;
export const Pm_resourceskillsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_resourceskillsstatuscode = keyof typeof Pm_resourceskillsstatuscode;

export interface Pm_resourceskillsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_certificationexpirydate?: string;
  pm_certificationname?: string;
  pm_certified?: boolean;
  pm_primaryskill?: boolean;
  pm_proficiencylevel?: Pm_resourceskillspm_proficiencylevel;
  "pm_resource@odata.bind"?: string;
  pm_resourceid?: string;
  pm_resourceskillid: string;
  "pm_skill@odata.bind"?: string;
  pm_skillid?: string;
  pm_yearsofexperience?: number;
  statecode: Pm_resourceskillsstatecode;
  statuscode?: Pm_resourceskillsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_resourceskills extends Pm_resourceskillsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_certifiedname?: string;
  pm_primaryskillname?: string;
  pm_proficiencylevelname?: string;
  pm_resourcename?: string;
  pm_skillname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_resource?: object;
  _pm_resource_value?: string;
  pm_skill?: object;
  _pm_skill_value?: string;
}